% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown2S3_addin.R
\name{rmarkdown2S3_Addin}
\alias{rmarkdown2S3_Addin}
\title{Publish rmarkdown report to S3 bucket}
\usage{
rmarkdown2S3_Addin()
}
\description{
Call this function to render the active document, which must be a \code{.rmd}
document, and then copy the resulting \code{.html} and any \code{_files/} to
an S3 bucket. The user will be asked to enter the name of the S3 bucket they
want to copy to.
}
